pyproject.toml
[tool.poetry]
name = "mstc-cloud-tools"
version = "0.1.0"
description = ""
authors = ["DELANNOYKOBAYASHI.MARCO.T.1568283135.E <marco.de_lannoy_kobayashi.ctr@us.af.mil>"]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.0.3"
requests = "^2.27.1"
Flask-API = "^3.0.post1"

[tool.poetry.dev-dependencies]
pytest = ">=6.0.0"
pytest_check = "^1.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
